+++ src/apps/nodejs.rs
@@ -2,7 +2,8 @@ use super::App;
 use crate::hosting::github_releases;
 use crate::install::archive::{self, InstallArgs};
 use crate::platform::{Cpu, Os, Platform};
-use crate::yard::{Executable, Yard};
+use crate::subshell::Executable;
+use crate::yard::Yard;
 use crate::{Output, Result};
 
 pub struct NodeJS {}
