+++ src/apps/shellcheck.rs
@@ -2,7 +2,8 @@ use super::App;
 use crate::hosting::github_releases;
 use crate::install::packaged_executable::{self, InstallArgs};
 use crate::platform::{Cpu, Os, Platform};
-use crate::yard::{Executable, Yard};
+use crate::subshell::Executable;
+use crate::yard::Yard;
 use crate::{Output, Result};
 
 pub struct ShellCheck {}
