+++ src/archives/zip.rs
@@ -1,7 +1,7 @@
 use super::Archive;
 use crate::filesystem::strip_filepath;
 use crate::output::Output;
-use crate::yard::Executable;
+use crate::subshell::Executable;
 use crate::{filesystem, Result};
 use std::path::Path;
 use std::{fs, io};
