+++ src/cmd/run.rs
@@ -5,8 +5,8 @@ use crate::error::UserError;
 use crate::filesystem::find_global_install;
 use crate::platform;
 use crate::subshell;
+use crate::subshell::Executable;
 use crate::yard;
-use crate::yard::Executable;
 use crate::Output;
 use crate::Result;
 use std::process::ExitCode;
