+++ src/install/compile_go.rs
@@ -1,6 +1,6 @@
 use crate::error::UserError;
 use crate::output::Output;
-use crate::yard::Executable;
+use crate::subshell::Executable;
 use crate::Result;
 use std::fs;
 use std::path::Path;
