+++ src/subshell/execute.rs
@@ -1,6 +1,6 @@
 use super::exit_status_to_code;
 use crate::error::UserError;
-use crate::yard::Executable;
+use crate::subshell::Executable;
 use crate::Result;
 use std::process::{Command, ExitCode};
 
@@ -24,7 +24,7 @@ mod tests {
         #[cfg(unix)]
         fn unix_success() {
             use crate::subshell::run;
-            use crate::yard::Executable;
+            use crate::subshell::Executable;
             use big_s::S;
             use std::io::Write;
             use std::os::unix::fs::PermissionsExt;
@@ -47,7 +47,7 @@ mod tests {
         fn unix_error() {
             use crate::filesystem::make_file_executable;
             use crate::subshell::run;
-            use crate::yard::Executable;
+            use crate::subshell::Executable;
             use big_s::S;
             use std::fs;
             let tempdir = tempfile::tempdir().unwrap();
@@ -64,7 +64,7 @@ mod tests {
         #[cfg(windows)]
         fn windows_success() {
             use crate::subshell::run;
-            use crate::yard::Executable;
+            use crate::subshell::Executable;
             use big_s::S;
             use std::fs;
             let tempdir = tempfile::tempdir().unwrap();
@@ -80,7 +80,7 @@ mod tests {
         #[cfg(windows)]
         fn windows_error() {
             use crate::subshell::run;
-            use crate::yard::Executable;
+            use crate::subshell::Executable;
             use big_s::S;
             use std::fs;
             let tempdir = tempfile::tempdir().unwrap();
