+++ src/subshell/stream.rs
@@ -1,6 +1,6 @@
 use super::exit_status_to_code;
 use crate::error::UserError;
-use crate::yard::Executable;
+use crate::subshell::Executable;
 use crate::Result;
 use std::io::{self, BufRead, BufReader, Read, Write};
 use std::process::{self, Child, Command, ExitCode, Stdio};
