+++ src/yard/mod.rs
@@ -2,7 +2,6 @@
 //! Named after rail yards, i.e. locations where passenger cars of trains are stored, sorted, and repaired.
 
 mod create;
-mod executable;
 mod load;
 mod load_or_create;
 #[allow(clippy::module_inception)] // I can't come up with a better name for this
@@ -10,7 +9,6 @@ mod yard;
 
 use crate::Result;
 pub use create::create;
-pub use executable::Executable;
 pub use load::load;
 pub use load_or_create::load_or_create;
 use std::path::{Path, PathBuf};
